Check if part touching virtual floor CanCollide
Greetings, I'm noticing a stutter in my character's flight when they touch a Pickup part which has CanCollide set to false.  With a little tracing I discovered that my character's state is changing from Freefall to Running when they touch the pickup.  I've made this little change locally and it has alleviated the issue.  Recommend to apply this patch upstream as well.

Best Regards,
[FirstVertex](https://devforum.roblox.com/u/firstvertex)